+++ packages/framework/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @manyullyn/framework
 
+## 0.0.4
+
+### Patch Changes
+
+- a1f8a96: Added APIChannel to CommandHandler
+- 0a8e538: Added deleteCommand
+- a3df2c3: Fix initializing commands property after using it
+- Added twurple API Provider and updated APIChannel
+- a1f8a96: Added Testing
+
 ## 0.0.3
 
 ### Patch Changes